// shipeasy/src/app/models/payment.ts
export interface Payment {
  paymentRefNo: string;
    batch_no: string;
    reciept_no: string;
    _id: { $oid: string };
    paymentTypeId: string;
    paymentType: string;
    billno: string;
    paymentreference_no: number;
    paymentDate: string;
    amount: number;
    remitance_bankId: string;
    remitance_bank: string;
    beneficiary_bankId: string;
    beneficiary_bank: string;
    invoice_partyId: string;
    invoice_party: string;
    recieved_from: string;
    recieved_fromId: string;
    batchNo: string;
    batchId: string;
    invoice_no: string;
    remarks: string;
    currencyId: string;
    currency: string;
    move_no: string;
    userType: string;
    module: string;
    isExport: boolean;
    payment_status: string;
    document_name: string;
    document_tag: string;
    cheque_amount: string;
    upload_document: string;
    filename: string;
    status: boolean;
    isDraft: boolean;
    tenantId: string;
    invoiceData: any[];
    onAccountValue: string;
    onAccountRemark: string;
    onAccountTdsAmt: string;
    onAccountYear: string;
    referenceId: string;
    paymentId: string;
    createdOn: string;
    updatedOn: string;
    createdBy: string;
    createdByUID: string;
    updatedBy: string;
    updatedByUID: string;
    __v: number;
  }
  