// shipeasy/src/app/models/add-agent-advise.ts
  interface Container {
    _id: string;
    tenantId: string;
    containerId: string;
    batchId: string;
    batchNo: string;
    mastercontainerId: string;
    containerNumber: string;
    containerTypeId: string;
    containerDescription: string;
    containerTypeName: string;
    containerType: string;
    containerSize: string;
    containerHeight: string;
    imoType: any;
    imoTypeId: any;
    netWeight: string;
    grossWeight: string;
    isoContainerCode: string;
    tareWeight: string;
    sealNo: string;
    unit: string;
    rfidNo: string;
    cargoTypeId: any;
    evgmNumber: string;
    evgmDate: string; // Or specify the actual type of the array elements
    blNumber: string;
    blDate: string;
    shippingBillNumber: string;
    sbNo: string;
    sbDate: string;
    bondNumber: string;
    igmNumber: string;
    statusFlagId: string;
    status: boolean;
    depotOut: string;
    depotDate: string | null;
    depotDateName: string;
    icdIn: string;
    icdInName: string;
    icdOut: string;
    icdOutName: string;
    factoryIn: string;
    factoryInName: string;
    factoryOut: string;
    factoryOutName: string;
    terminalIn: string;
    terminalInName: string;
    terminalOut: string;
    terminalOutName: string;
    mtyValidity: string;
    mtyReturn: string;
    cfsIn: string;
    cfsOut: string;
    railOut: string;
    dischargeDate: string;
    reject: string;
    rejectName: string;
    sobDate: string;
    arrivalDate: string;
    deliveryDate: string;
    override_orgId: string;
    override_tId: string;
    containerInUse: boolean;
    isExport: boolean;
    referenceId: string;
    createdOn: string;
    updatedOn: string;
    __v: number;
  }
  
  interface ShippingDocument {
    // Define properties for shipping documents if needed
  }
  
 export interface Bl {
  seqCharacter: string; 
    pdfUrl: string;
    tenantId: string;
    blId: string;
    batchId: string;
    Import_Export: string;
    blType: string;
    subBltype: string;
    blTypeName: string;
    blNumber: string;
    shipperId: string;
    shipperName: string;
    shipperAddress: string;
    consigneeId: string;
    consigneeName: string;
    consigneeAddress: string;
    voyageId: string;
    vessel: string;
    vesselName: string;
    preCarrigeById: string;
    preCarrigeByName: string;
    notify_party1: string;
    address1: string;
    notify_party2: string;
    address2: string;
    notify_party3: string;
    address3: string;
    loadPlace: string;
    loadPlaceName: string;
    entryPort: string;
    entryPortId: string;
    onCarriageId: string;
    onCarriageName: string;
    placeOfDelivery: string;
    stoltAgentId: string;
    stoltAgentName: string;
    manifestRemarks: boolean;
    podUnLoc: string;
    polId: string;
    polName: string;
    placeofIssue: string;
    placeofReceipt: string;
    frieghtPaidBy: string;
    blINCOTerm: string;
    billToId: string;
    billToName: string;
    doInvoice: boolean;
    documentInvoice: boolean;
    handlingFees: string;
    pcin: string;
    csn: string;
    mcin: string;
    blStatus: boolean;
    additional: string;
    containers: Container[];
    status: boolean;
    shippingLineId: string;
    mbl: string;
    IGM_Filed: string;
    blDate: string;
    shippingTermId: string;
    isDPD: boolean;
    isHSS: boolean;
    grossWeight: string;
    nettWeight: string;
    markNumber: string;
    totalPackage: string;
    cargo_Desc: string;
    cargoType: string;
    cargoId: string;
    cfsLocationId: string;
    cargoStatus: string;
    unNo: string;
    imcoClass: string;
    technicalName: string;
    productId: string;
    surveyor: string;
    emptyReturnDepot: string;
    VesselSellingDate: string;
    LumSumpDatefrom: string;
    LumSumpDateTo: string;
    LumSumpDaysinPeriod: string;
    amount: string;
    noofContainer: string;
    importPoo: string;
    importPol: string;
    importPod: string;
    importFpod: string;
    departureMode: string;
    blGeneralRemarks: string;
    documents: ShippingDocument[];
    polRemarks: string;
    containerType: string;
    itemNo: string;
    subLineNo: string;
    isExport: boolean;
    apiType: string;
    referenceId: string;
    createdOn: string;
    updatedOn: string;
    createdBy: string;
    createdByUID: string;
    updatedBy: string;
    updatedByUID: string;
    __v: number;
  }